Add test coverage
Using [nyc][nyc] and [coveralls.io][coveralls].

[nyc]: https://github.com/istanbuljs/nyc
[coveralls]: https://coveralls.io/github/runarberg/formsquare
