+++ .gitignore
@@ -2,3 +2,5 @@ dist/
 lib/
 node_modules/
 index.js
+.nyc_output/
+coverage/
